Move SandboxHelper.cs up out of resources. Add readme to resources.
